Move Record.Like to Graph and tidy up
