+++ src/idunno.Bluesky/Feed/Model/GetLikesResponse.cs
@@ -1,19 +1,15 @@
 // Copyright (c) Barry Dorrans. All rights reserved.
 // Licensed under the MIT License.
 
-using System.Diagnostics.CodeAnalysis;
 using System.Text.Json.Serialization;
 
 using idunno.AtProto;
-using idunno.Bluesky.Feed.Likes;
-
 namespace idunno.Bluesky.Feed.Model;
 
-[SuppressMessage("Performance", "CA1812", Justification = "Used in GetLikes.")]
 internal sealed record GetLikesResponse
 {
     [JsonConstructor]
-    public GetLikesResponse(AtUri uri, Cid? cid, ICollection<Like> likes, string? cursor)
+    public GetLikesResponse(AtUri uri, Cid? cid, ICollection<Likes.Like> likes, string? cursor)
     {
         Uri = uri;
         Cid = cid;
@@ -30,7 +26,7 @@ internal sealed record GetLikesResponse
 
     [JsonInclude]
     [JsonRequired]
-    public ICollection<Like> Likes { get; init; }
+    public ICollection<Likes.Like> Likes { get; init; }
 
     [JsonInclude]
     public string? Cursor { get; init; }
