+++ src/idunno.Bluesky/Record/BlueskyRecord.cs
@@ -16,7 +16,7 @@ namespace idunno.Bluesky.Record;
 [JsonDerivedType(typeof(Post), RecordType.Post)]
 [JsonDerivedType(typeof(Follow), RecordType.Follow)]
 [JsonDerivedType(typeof(Repost), RecordType.Repost)]
-[JsonDerivedType(typeof(Like), RecordType.Like)]
+[JsonDerivedType(typeof(Feed.Like), RecordType.Like)]
 [JsonDerivedType(typeof(Block), RecordType.Block)]
 [JsonDerivedType(typeof(Profile), RecordType.Profile)]
 [JsonDerivedType(typeof(StarterPack), RecordType.StarterPack)]
